Remove last dependencies on `static.mjs`
