+++ production/skilljar-theme-v3.0/course-completion.mjs
@@ -1,5 +1,5 @@
 import { el, text, Q } from "./meta.mjs";
-import { CONFIG } from "./static.mjs";
+import { CONFIG } from "./config.mjs";
 import { setStyle } from "./styling.mjs";
 import { CG } from "./CG.mjs";
