--- production/skilljar-theme-v3.0/static.mjs
@@ -1,64 +1,0 @@
-import { el } from "./meta.mjs";
-
-export const CONFIG = {
-  logo: el(
-    "svg",
-    {
-      xmlns: "http://www.w3.org/2000/svg",
-      viewBox: "38 45 402 41",
-      fill: "#6226FB",
-      class: "header-center-img header-logo-svg loading",
-      alt: "Go home",
-      "data-was-processed": "true",
-    },
-    [
-      el("path", {
-        "fill-rule": "evenodd",
-        "clip-rule": "evenodd",
-        d: "M73.1585 67.6636C73.4859 66.5314 73.6637 65.2907 73.6637 63.9412C73.6637 55.5005 66.7016 46 59.748 46C52.7944 46 45.8322 55.5005 45.8322 63.9412C45.8322 65.7815 46.1631 67.4196 46.7528 68.856L42.573 68.6204C40.8706 68.5244 39.3022 69.7413 39.5677 71.4298C39.6707 72.0848 39.8669 72.7552 40.2175 73.3341C39.0342 74.1686 38.5508 75.7027 39.5096 76.905C40.4721 78.1118 41.8853 79.2239 43.8251 79.2239C45.9291 79.2239 47.1954 78.6787 47.9644 77.9941C48.0205 78.2931 48.1363 78.5875 48.3191 78.8657C49.2634 80.3031 50.8274 81.883 53.1396 81.883C57.073 81.883 57.5116 79.4063 57.7318 78.1635C57.749 78.0662 57.7649 77.9765 57.781 77.8961L60.4095 76.5782L63.038 77.8961C63.0541 77.9765 63.07 78.066 63.0871 78.1633C63.3074 79.4061 63.7459 81.883 67.6794 81.883C69.9914 81.883 71.5556 80.3031 72.4999 78.8657C72.537 78.8093 72.5714 78.7521 72.6029 78.6942C73.3441 79.0154 74.325 79.2239 75.6263 79.2239C77.5662 79.2239 78.9794 78.1118 79.942 76.905C81.1207 75.4266 80.1194 73.4471 78.3174 72.8826L77.7221 72.6961C79.4704 71.7875 79.7277 70.0743 79.5405 68.573C79.3289 66.8766 77.4855 66.1434 75.8764 66.7087L73.1585 67.6636ZM60.412 76.5539H60.4073L60.4095 76.5551L60.412 76.5539Z",
-      }),
-      el("path", {
-        d: "M104.707 78.1485C96.7215 78.1485 91.6155 72.9526 91.6155 63.7923C91.6155 55.1324 97.3741 49.821 104.975 49.821C111.54 49.821 116.262 53.5929 116.954 59.9819H111.732C111.233 56.4795 108.93 54.0932 104.975 54.0932C100.215 54.0932 96.7983 57.7496 96.7983 63.7923C96.7983 70.0659 100.177 73.8764 104.86 73.8764C108.853 73.8764 111.502 71.2976 112.039 67.7181H117.069C116.416 73.9534 111.924 78.1485 104.707 78.1485ZM120.729 77.4557V50.5138H125.413V59.5201C125.413 60.0589 125.375 60.6363 125.259 61.3675C126.296 59.4816 128.062 58.0575 131.018 58.0575C135.164 58.0575 137.621 60.9827 137.621 65.6013V77.4557H132.937V66.2941C132.937 63.5229 131.594 61.7909 129.406 61.7909C127.102 61.7909 125.413 63.8308 125.413 66.448V77.4557H120.729ZM147.701 78.1485C144.437 78.1485 141.251 75.993 141.251 72.2984C141.251 68.0645 144.514 66.602 148.545 66.0631L151.233 65.7167C152.768 65.5243 153.306 64.947 153.306 64.0233C153.306 62.7146 152.231 61.56 150.234 61.56C148.046 61.56 146.664 62.7916 146.472 64.7545H141.635C141.942 60.8287 145.128 58.0575 149.966 58.0575C155.686 58.0575 157.989 61.1366 157.989 66.4095V77.4557H153.651V76.6859C153.651 76.07 153.728 75.4928 153.843 74.8769C152.845 76.7245 150.925 78.1485 147.701 78.1485ZM148.967 74.6846C151.463 74.6846 153.459 72.837 153.459 70.1815V68.1415C153.037 68.5264 152.384 68.7188 151.233 68.9498L149.697 69.2577C147.508 69.681 146.011 70.4124 146.011 72.1058C146.011 73.7608 147.317 74.6846 148.967 74.6846ZM162.569 77.4557V58.6733H167.253V77.4557H162.569ZM162.454 55.8252V50.8987H167.33V55.8252H162.454ZM172.092 77.4557V58.6733H176.776V59.5201C176.776 60.0589 176.737 60.6363 176.622 61.3675C177.659 59.4816 179.424 58.0575 182.381 58.0575C186.527 58.0575 188.984 60.9827 188.984 65.6013V77.4557H184.3V66.2941C184.3 63.5229 182.956 61.7909 180.768 61.7909C178.465 61.7909 176.776 63.8308 176.776 66.448V77.4557H172.092ZM201.981 86C196.453 86 193.305 83.0365 193.305 78.7644H198.065C198.142 80.9966 199.562 82.3053 202.096 82.3053C204.553 82.3053 206.626 80.8427 206.626 78.0329V75.993C206.626 75.1465 206.664 74.3766 206.78 73.6454C205.666 75.3388 203.862 76.7629 200.944 76.7629C196.069 76.7629 192.537 73.3375 192.537 67.4102C192.537 61.56 196.606 58.0575 200.714 58.0575C204.015 58.0575 205.666 59.4046 206.78 61.3675C206.664 60.7132 206.626 60.2129 206.626 59.5201V58.6733H211.31V76.8784C211.31 83.0365 207.509 86 201.981 86ZM202.134 73.0296C204.898 73.0296 206.895 70.6817 206.895 67.4102C206.895 64.1387 204.898 61.7909 202.134 61.7909C199.293 61.7909 197.297 64.1387 197.297 67.4102C197.297 70.6817 199.293 73.0296 202.134 73.0296ZM222.633 78.1485C218.87 78.1485 215.914 75.5697 215.914 70.6048V58.6733H220.636V69.8735C220.636 72.4907 221.788 74.3766 224.053 74.3766C226.433 74.3766 228.122 72.606 228.122 69.8349V58.6733H232.806V77.4557H228.161V76.6859C228.161 76.1086 228.238 75.4544 228.353 74.8769C227.393 76.6859 225.742 78.1485 222.633 78.1485ZM243.077 78.1485C239.814 78.1485 236.628 75.993 236.628 72.2984C236.628 68.0645 239.891 66.602 243.922 66.0631L246.609 65.7167C248.145 65.5243 248.683 64.947 248.683 64.0233C248.683 62.7146 247.608 61.56 245.611 61.56C243.423 61.56 242.041 62.7916 241.849 64.7545H237.012C237.319 60.8287 240.505 58.0575 245.342 58.0575C251.063 58.0575 253.366 61.1366 253.366 66.4095V77.4557H249.028V76.6859C249.028 76.07 249.105 75.4928 249.22 74.8769C248.222 76.7245 246.302 78.1485 243.077 78.1485ZM244.344 74.6846C246.84 74.6846 248.836 72.837 248.836 70.1815V68.1415C248.414 68.5264 247.761 68.7188 246.609 68.9498L245.074 69.2577C242.886 69.681 241.388 70.4124 241.388 72.1058C241.388 73.7608 242.694 74.6846 244.344 74.6846ZM257.946 77.4557V58.6733H262.63V60.0974C262.63 60.7902 262.591 61.3675 262.476 62.0603C263.474 59.9435 265.164 58.0575 267.966 58.0575C268.35 58.0575 268.657 58.096 268.964 58.173V62.7146C268.657 62.6377 268.235 62.5607 267.621 62.5607C264.741 62.5607 262.63 64.0233 262.63 68.1415V77.4557H257.946ZM278.892 78.1485C273.939 78.1485 270.33 74.7616 270.33 68.1415C270.33 61.406 274.592 58.0575 278.968 58.0575C281.886 58.0575 283.689 59.5586 284.765 61.3675C284.611 60.6747 284.611 60.0974 284.611 59.5201V50.5138H289.295V77.4557H284.611V76.6859C284.611 76.0316 284.611 75.5314 284.765 74.8769C284.036 76.2626 282.346 78.1485 278.892 78.1485ZM280.004 74.4536C282.808 74.4536 284.881 71.8748 284.881 68.103C284.881 64.2542 282.808 61.7909 280.004 61.7909C277.202 61.7909 275.091 64.3696 275.091 68.103C275.091 71.9904 277.202 74.4536 280.004 74.4536Z",
-      }),
-      el("path", {
-        d: "M315.566 78.3906C312.897 78.3906 310.475 77.7656 308.301 76.5156C306.139 75.2656 304.44 73.5404 303.203 71.3398C301.979 69.1393 301.367 66.6914 301.367 63.9961C301.367 61.3008 301.979 58.8529 303.203 56.6523C304.44 54.4518 306.139 52.7266 308.301 51.4766C310.475 50.2266 312.897 49.6016 315.566 49.6016C317.741 49.6016 319.798 50.0443 321.738 50.9297C323.678 51.8021 325.286 53.026 326.562 54.6016C327.852 56.1641 328.626 57.9089 328.887 59.8359H322.285C321.908 58.5208 321.094 57.4466 319.844 56.6133C318.607 55.7799 317.181 55.3633 315.566 55.3633C313.262 55.3633 311.328 56.1901 309.766 57.8438C308.216 59.4844 307.441 61.5352 307.441 63.9961C307.441 66.457 308.216 68.5143 309.766 70.168C311.328 71.8216 313.262 72.6484 315.566 72.6484C317.194 72.6484 318.633 72.1927 319.883 71.2812C321.146 70.3698 321.96 69.2109 322.324 67.8047H328.965C328.431 70.9167 326.914 73.4622 324.414 75.4414C321.927 77.4076 318.978 78.3906 315.566 78.3906ZM339.845 57.9609C341.265 57.9609 342.599 58.2214 343.849 58.7422C345.112 59.263 346.193 59.9792 347.091 60.8906C348.003 61.7891 348.719 62.8763 349.24 64.1523C349.774 65.4154 350.041 66.763 350.041 68.1953C350.041 70.1094 349.591 71.8477 348.693 73.4102C347.808 74.9727 346.584 76.1966 345.021 77.082C343.472 77.9544 341.746 78.3906 339.845 78.3906C336.942 78.3906 334.513 77.4206 332.56 75.4805C330.607 73.5404 329.63 71.112 329.63 68.1953C329.63 65.2786 330.607 62.8438 332.56 60.8906C334.513 58.9375 336.942 57.9609 339.845 57.9609ZM339.845 63.3516C338.504 63.3516 337.384 63.8138 336.486 64.7383C335.601 65.6497 335.158 66.8021 335.158 68.1953C335.158 69.5495 335.601 70.6888 336.486 71.6133C337.384 72.5378 338.504 73 339.845 73C341.16 73 342.267 72.5378 343.166 71.6133C344.064 70.6888 344.513 69.5495 344.513 68.1953C344.513 66.8021 344.064 65.6497 343.166 64.7383C342.28 63.8138 341.173 63.3516 339.845 63.3516ZM365.003 58.3516H370.648V78H365.003V75.4023C363.766 77.3945 361.995 78.3906 359.691 78.3906C357.178 78.3906 355.198 77.5964 353.753 76.0078C352.321 74.4193 351.605 72.2318 351.605 69.4453V58.3516H357.249V68.7227C357.249 70.0247 357.601 71.0664 358.304 71.8477C359.007 72.6159 359.951 73 361.136 73C362.295 73 363.226 72.6159 363.929 71.8477C364.645 71.0664 365.003 70.0247 365.003 68.7227V58.3516ZM383.735 57.9609C384.139 57.9609 384.77 58.0391 385.63 58.1953V63.3125C384.835 63.1302 384.152 63.0391 383.579 63.0391C382.134 63.0391 380.949 63.4883 380.024 64.3867C379.113 65.2721 378.657 66.5807 378.657 68.3125V78H373.012V58.3516H378.657V60.5977C379.868 58.8398 381.561 57.9609 383.735 57.9609ZM394.108 78.3906C391.517 78.3906 389.401 77.7721 387.76 76.5352C386.133 75.2852 385.221 73.5599 385.026 71.3594H390.436C390.488 72.0234 390.833 72.5573 391.471 72.9609C392.109 73.3646 392.897 73.5664 393.834 73.5664C394.668 73.5664 395.312 73.4362 395.768 73.1758C396.237 72.9154 396.471 72.5508 396.471 72.082C396.471 71.6523 396.269 71.3008 395.866 71.0273C395.462 70.7539 394.935 70.5521 394.284 70.4219C393.633 70.2786 392.91 70.1419 392.116 70.0117C391.334 69.8815 390.553 69.6992 389.772 69.4648C388.991 69.2305 388.274 68.931 387.623 68.5664C386.972 68.2018 386.445 67.668 386.041 66.9648C385.638 66.2487 385.436 65.3958 385.436 64.4062C385.436 62.5443 386.139 61.0078 387.545 59.7969C388.965 58.5729 390.859 57.9609 393.229 57.9609C395.468 57.9609 397.389 58.5339 398.991 59.6797C400.592 60.8255 401.51 62.4401 401.745 64.5234H396.061C396.022 64.0026 395.735 63.5599 395.202 63.1953C394.668 62.8307 393.971 62.6484 393.112 62.6484C392.435 62.6484 391.888 62.7656 391.471 63C391.067 63.2214 390.866 63.5534 390.866 63.9961C390.866 64.3997 391.067 64.7253 391.471 64.9727C391.888 65.207 392.422 65.3828 393.073 65.5C393.724 65.6172 394.44 65.7344 395.221 65.8516C396.002 65.9688 396.784 66.151 397.565 66.3984C398.359 66.6458 399.082 66.9779 399.733 67.3945C400.384 67.7982 400.911 68.3971 401.315 69.1914C401.732 69.9857 401.94 70.9362 401.94 72.043C401.94 73.8398 401.185 75.3503 399.674 76.5742C398.177 77.7852 396.321 78.3906 394.108 78.3906ZM422.098 67.5117C422.098 68.2018 422.059 68.8724 421.98 69.5234H407.391C407.586 70.7995 408.139 71.8216 409.051 72.5898C409.962 73.3451 411.076 73.7227 412.391 73.7227C413.471 73.7227 414.35 73.5143 415.027 73.0977C415.717 72.668 416.141 72.1016 416.297 71.3984H421.824C421.499 73.5078 420.49 75.2005 418.797 76.4766C417.104 77.7526 415.021 78.3906 412.547 78.3906C410.646 78.3906 408.901 77.9414 407.312 77.043C405.737 76.1315 404.493 74.888 403.582 73.3125C402.671 71.724 402.215 69.9792 402.215 68.0781C402.215 65.2266 403.198 62.8307 405.164 60.8906C407.143 58.9375 409.565 57.9609 412.43 57.9609C414.174 57.9609 415.789 58.3776 417.273 59.2109C418.758 60.0443 419.93 61.1966 420.789 62.668C421.661 64.1393 422.098 65.7539 422.098 67.5117ZM407.664 65.9297H416.219C416.141 64.9531 415.711 64.1393 414.93 63.4883C414.148 62.8372 413.23 62.5117 412.176 62.5117C411.069 62.5117 410.125 62.7982 409.344 63.3711C408.562 63.944 408.003 64.7969 407.664 65.9297ZM431.455 78.3906C428.864 78.3906 426.748 77.7721 425.107 76.5352C423.479 75.2852 422.568 73.5599 422.373 71.3594H427.783C427.835 72.0234 428.18 72.5573 428.818 72.9609C429.456 73.3646 430.244 73.5664 431.181 73.5664C432.015 73.5664 432.659 73.4362 433.115 73.1758C433.584 72.9154 433.818 72.5508 433.818 72.082C433.818 71.6523 433.616 71.3008 433.213 71.0273C432.809 70.7539 432.282 70.5521 431.63 70.4219C430.979 70.2786 430.257 70.1419 429.463 70.0117C428.681 69.8815 427.9 69.6992 427.119 69.4648C426.338 69.2305 425.621 68.931 424.97 68.5664C424.319 68.2018 423.792 67.668 423.388 66.9648C422.985 66.2487 422.783 65.3958 422.783 64.4062C422.783 62.5443 423.486 61.0078 424.892 59.7969C426.311 58.5729 428.206 57.9609 430.576 57.9609C432.815 57.9609 434.736 58.5339 436.338 59.6797C437.939 60.8255 438.857 62.4401 439.091 64.5234H433.408C433.369 64.0026 433.082 63.5599 432.548 63.1953C432.015 62.8307 431.318 62.6484 430.459 62.6484C429.782 62.6484 429.235 62.7656 428.818 63C428.414 63.2214 428.213 63.5534 428.213 63.9961C428.213 64.3997 428.414 64.7253 428.818 64.9727C429.235 65.207 429.768 65.3828 430.42 65.5C431.071 65.6172 431.787 65.7344 432.568 65.8516C433.349 65.9688 434.13 66.151 434.912 66.3984C435.706 66.6458 436.429 66.9779 437.08 67.3945C437.731 67.7982 438.258 68.3971 438.662 69.1914C439.078 69.9857 439.287 70.9362 439.287 72.043C439.287 73.8398 438.532 75.3503 437.021 76.5742C435.524 77.7852 433.668 78.3906 431.455 78.3906Z",
-      }),
-    ],
-  ),
-  domains: {
-    prod: { url: "courses.chainguard.dev", id: "3glgawqmzatte" },
-    stage: { url: "chainguard-test.skilljar.com", id: "ix1ljpxex6xd" },
-  },
-  partners: {
-    "chainguard-discovery-partner-sales-foundations": {
-      id: "53njmyk25y3v",
-      checkout: "19bn1isfg4c3t",
-    },
-    "chainguard-advanced-partner-sales-accelerator": {
-      id: "1w57muf27zdg1",
-      checkout: "3em1yw57v5d30",
-    },
-  },
-  confetti: {
-    autoHideMs: 6000,
-    particles: {
-      stars: { count: 40, scalar: 1.2 },
-      circles: { count: 10, scalar: 0.75 },
-      logos: { count: 50, scalar: 3.0 },
-    },
-    defaults: {
-      spread: 360,
-      ticks: 50,
-      gravity: 1,
-      decay: 0.94,
-      startVelocity: 40,
-      shapes: ["star"],
-      colors: ["#C6FF50", "#50FFE1"],
-    },
-  },
-  codeTheme: "min-light", // Shiki theme to use for code blocks
-};
-
-export { icons, bannerSVGs } from "../data/graphics.mjs";
-export { footerData } from "../data/footer.mjs";
-export { pathSections } from "../data/path-sections.mjs";
